requirements.txt
streamlit
openai
fpdf2
requests
youtube-transcript-api
pandas
